kpps: drop no-op KeepAlive calls and tidy the tests
Three more review comments on the kernel PPS source package.

The runtime.KeepAlive calls in ioctlGetCap and ioctlFetch do nothing: ioctlPtr converts the unsafe.Pointer to uintptr inside the Syscall argument list, which is the pattern the compiler recognizes to keep the object alive for the call. term's ioctlPtr has no such calls either.

TestCaptureModes compared the two mode constants with the literals they are declared with a few lines away, so it could not fail; x/sys/unix has no PPS mode constants to check them against, so the test goes and types_test.go with it. TestInfoFromKernel moves to ioctl_linux_test.go, next to the function it tests.
